[Refactor] 라우터 리팩토링
