+++ client/src/App.jsx
@@ -43,7 +43,7 @@ function App() {
             }
           />
           <Route
-            path="/meetinglists/task/:id"
+            path="/meetingData/meetinglists/task/:id"
             element={
               <SidebarLayout>
                 <MeetingList />
